feat: 기본 기능 서비스 개발 & 단위테스트 작성 중

+++ src/main/java/io/hhplus/server/controller/user/dto/response/GetBalanceResponse.java
@@ -1,6 +1,8 @@
 package io.hhplus.server.controller.user.dto.response;
 
+import java.math.BigDecimal;
+
 public record GetBalanceResponse(
-        int balance
+        BigDecimal balance
 ) {
 }
