+++ src/main/java/io/hhplus/server/controller/user/dto/response/GetMyReservationsResponse.java
@@ -4,6 +4,7 @@ import io.hhplus.server.domain.payment.PaymentEnums;
 import io.hhplus.server.domain.reservation.ReservationEnums;
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 
 public record GetMyReservationsResponse(
@@ -32,7 +33,7 @@ public record GetMyReservationsResponse(
     public static record PaymentInfo(
             Long paymentId,
             PaymentEnums.Status status,
-            int paymentPrice
+            BigDecimal paymentPrice
     ) {
     }
